♻️ refactor/ZIP-44 : 시작후, 바로 배치처리 및 DTO 수정
